Consume shared YOLO Swift package on iOS
Replaces the vendored, diverged copy of the iOS inference core with a dependency
on the YOLO Swift package (ultralytics/yolo-ios-app), so the plugin and the
native iOS app share one source of truth.

- Delete the duplicated core from ios/Classes (BasePredictor, all predictors,
  Predictor, NonMaxSuppression, ThresholdProvider, YOLOResult, YOLOTask, Plot,
  UIImage+Orientation, YOLO); these now come from `import YOLO`.
- Keep the Flutter-specific view/camera/bridge layer (YOLOView, VideoCapture,
  YOLOCamera, BoundingBoxView, YOLOPlugin, YOLOInstanceManager,
  SwiftYOLOPlatformView(+Factory), YOLOStreamConfig). Add YOLOOverlayStyle for
  the live-overlay drawing previously provided by Plot (view-layer, per-consumer;
  palette/skeleton are an exact copy).
- Remap the former predictor streamConfig coupling: original-image capture maps
  to the package's BasePredictor.capturesOriginalImage; all other stream
  filtering stays in the plugin's result-serialization path.
- ultralytics_yolo.podspec depends on YOLO ~> 0.1.

Release ordering: the example Podfile temporarily resolves YOLO from the
yolo-ios-app branch. Do not publish to pub.dev until YOLO 0.1.0 is tagged and
pushed to CocoaPods trunk; then drop the Podfile override.

Gates: example iOS build (CocoaPods) green, flutter analyze clean, 155 Dart
tests pass.
