--- ios/Classes/NonMaxSuppression.swift
@@ -1,58 +1,0 @@
-// Ultralytics 🚀 AGPL-3.0 License - https://ultralytics.com/license
-
-//
-//  This file is part of the Ultralytics YOLO Package, providing utilities for post-processing detections.
-//  Licensed under AGPL-3.0. For commercial use, refer to Ultralytics licensing: https://ultralytics.com/license
-//  Access the source code: https://github.com/ultralytics/yolo-ios-app
-//
-//  The NonMaxSuppression utility provides a fundamental algorithm for filtering redundant
-//  object detections. When multiple bounding boxes detect the same object, non-maximum suppression
-//  selects the best detection and suppresses overlapping boxes with lower confidence scores.
-//  This implementation sorts detection candidates by confidence score, then iteratively selects
-//  the highest-scoring boxes while removing others that have significant overlap (measured by
-//  intersection over minimum area). The algorithm is essential for producing clean detection
-//  results by removing duplicate predictions.
-
-import Foundation
-
-/// Performs non-maximum suppression on a set of bounding boxes to eliminate duplicate detections.
-/// - Parameters:
-///   - boxes: Array of bounding boxes in CGRect format.
-///   - scores: Confidence scores corresponding to each bounding box.
-///   - threshold: The minimum overlap ratio required to suppress a box.
-/// - Returns: Indices of the selected bounding boxes after suppression.
-public func nonMaxSuppression(boxes: [CGRect], scores: [Float], threshold: Float) -> [Int] {
-  // Suppress by true IoU (intersection-over-UNION), matching yolo-ios-app. The previous code used
-  // intersection-over-MIN-area, which over-suppresses (a small box fully inside a larger one always exceeds the
-  // threshold) and drops valid nested/nearby detections.
-  let count = boxes.count
-  let sortedIndices = scores.enumerated().sorted { $0.element > $1.element }.map { $0.offset }
-  let areas = boxes.map { $0.area }
-  var selectedIndices = [Int]()
-  var activeIndices = [Bool](repeating: true, count: count)
-  let iouThreshold = CGFloat(threshold)
-
-  for i in 0..<sortedIndices.count {
-    let idx = sortedIndices[i]
-    if !activeIndices[idx] { continue }
-    selectedIndices.append(idx)
-    let boxA = boxes[idx]
-    let areaA = areas[idx]
-    for j in (i + 1)..<sortedIndices.count {
-      let otherIdx = sortedIndices[j]
-      if !activeIndices[otherIdx] { continue }
-      let interArea = boxA.intersection(boxes[otherIdx]).area
-      let union = areaA + areas[otherIdx] - interArea
-      if union > 0 && interArea / union > iouThreshold {
-        activeIndices[otherIdx] = false
-      }
-    }
-  }
-  return selectedIndices
-}
-
-extension CGRect {
-  var area: CGFloat {
-    return width * height
-  }
-}
