--- ios/Classes/ObjectDetector.swift
@@ -1,274 +1,0 @@
-// Ultralytics 🚀 AGPL-3.0 License - https://ultralytics.com/license
-
-//
-//  This file is part of the Ultralytics YOLO Package, implementing object detection functionality.
-//  Licensed under AGPL-3.0. For commercial use, refer to Ultralytics licensing: https://ultralytics.com/license
-//  Access the source code: https://github.com/ultralytics/yolo-ios-app
-//
-//  The ObjectDetector class provides specialized functionality for detecting objects in images
-//  using YOLO models. It processes Vision framework results to extract bounding boxes, class labels,
-//  and confidence scores from model predictions. The class handles both real-time frame processing
-//  and single image analysis, converting the Vision API's normalized coordinates to image coordinates,
-//  and packaging the results in the standardized YOLOResult format. It includes performance monitoring
-//  for inference time and frame rate, and offers runtime adjustable parameters such as confidence
-//  threshold and IoU threshold for non-maximum suppression. It supports both the traditional Vision
-//  NMS pipeline (YOLO11, returning VNRecognizedObjectObservation) and NMS-free end2end models (YOLO26,
-//  returning raw MLMultiArray tensors), as well as traditional raw tensors decoded with Swift-side NMS.
-
-import CoreML
-import Foundation
-import UIKit
-import Vision
-
-/// Specialized predictor for YOLO object detection models that identifies and localizes objects in images.
-///
-/// This class processes the outputs from YOLO object detection models, extracting bounding boxes,
-/// class labels, and confidence scores. It handles both real-time camera feed processing and
-/// single image analysis, converting the normalized coordinates from the Vision framework
-/// to image coordinates and applying non-maximum suppression to filter duplicative detections.
-///
-/// - Note: Object detection models output rectangular bounding boxes around detected objects.
-/// - SeeAlso: `Segmenter` for models that produce pixel-level masks for objects.
-class ObjectDetector: BasePredictor, @unchecked Sendable {
-
-  /// Sets the confidence threshold and updates the model's feature provider.
-  ///
-  /// This overridden method ensures that when the confidence threshold is changed,
-  /// the Vision model's feature provider is also updated to use the new value.
-  ///
-  /// - Parameter confidence: The new confidence threshold value (0.0 to 1.0).
-  override func setConfidenceThreshold(confidence: Double) {
-    confidenceThreshold = confidence
-    // Force IoU = 1.0 for NMS-free (YOLO26) models so Vision's NMS doesn't suppress the model's own decoding — same as
-    // the create()-time seed, which this setter would otherwise clobber on the first threshold update.
-    detector.featureProvider = ThresholdProvider(
-      iouThreshold: requiresNMS ? iouThreshold : 1.0, confidenceThreshold: confidenceThreshold)
-  }
-
-  /// Sets the IoU threshold and updates the model's feature provider.
-  ///
-  /// This overridden method ensures that when the IoU threshold is changed,
-  /// the Vision model's feature provider is also updated to use the new value.
-  ///
-  /// - Parameter iou: The new IoU threshold value (0.0 to 1.0).
-  override func setIouThreshold(iou: Double) {
-    iouThreshold = iou
-    detector.featureProvider = ThresholdProvider(
-      iouThreshold: requiresNMS ? iouThreshold : 1.0, confidenceThreshold: confidenceThreshold)
-  }
-
-  /// Processes the results from the Vision framework's object detection request.
-  ///
-  /// This method extracts bounding boxes, class labels, and confidence scores from the
-  /// Vision object detection results, converts coordinates to the original image space,
-  /// and notifies listeners with the structured detection results.
-  ///
-  /// - Parameters:
-  ///   - request: The completed Vision request containing object detection results.
-  ///   - error: Any error that occurred during the Vision request.
-  override func processObservations(for request: VNRequest, error: Error?) {
-    let boxes = decodeBoxes(from: request)
-    let timing = updateTiming()
-    var result = YOLOResult(
-      orig_shape: inputSize, boxes: boxes, speed: timing.speed, fps: timing.fps, names: labels)
-
-    if let originalImageData = self.originalImageData {
-      result.originalImage = UIImage(data: originalImageData)
-    }
-
-    self.currentOnResultsListener?.on(result: result)
-  }
-
-  /// Processes a static image and returns object detection results.
-  ///
-  /// This method performs object detection on a static image and returns the
-  /// detection results synchronously. It handles the entire inference pipeline
-  /// from setting up the Vision request to processing the detection results.
-  ///
-  /// - Parameter image: The CIImage to analyze for object detection.
-  /// - Returns: A YOLOResult containing the detected objects with bounding boxes, class labels, and confidence scores.
-  override func predictOnImage(image: CIImage) -> YOLOResult {
-    guard let request = visionRequest else {
-      let emptyResult = YOLOResult(orig_shape: inputSize, boxes: [], speed: 0, names: labels)
-      return emptyResult
-    }
-    var boxes = [Box]()
-
-    let requestHandler = makeRequestHandler(for: image)
-    if perform(request, with: requestHandler, errorMessage: "YOLO ObjectDetector error") {
-      boxes = decodeBoxes(from: request)
-    }
-    var result = YOLOResult(orig_shape: inputSize, boxes: boxes, speed: 0, names: labels)
-    let annotatedImage = drawYOLODetections(on: image, result: result)
-    result.speed = finishTiming(notify: false)
-    result.annotatedImage = annotatedImage
-
-    return result
-  }
-
-  /// Decodes detection boxes from a completed Vision request.
-  ///
-  /// Handles both NMS-pipelined models (`VNRecognizedObjectObservation`, e.g. YOLO11) and NMS-free models
-  /// (`VNCoreMLFeatureValueObservation` raw tensors, e.g. YOLO26).
-  private func decodeBoxes(from request: VNRequest) -> [Box] {
-    if let results = request.results as? [VNRecognizedObjectObservation] {
-      return decodeRecognizedBoxes(results)
-    }
-    if let results = request.results as? [VNCoreMLFeatureValueObservation],
-      let prediction = results.first?.featureValue.multiArrayValue
-    {
-      return processRawResults(prediction)
-    }
-    return []
-  }
-
-  /// Decodes boxes from Vision's built-in NMS pipeline (`VNRecognizedObjectObservation`, e.g. YOLO11).
-  private func decodeRecognizedBoxes(_ results: [VNRecognizedObjectObservation]) -> [Box] {
-    var boxes = [Box]()
-    boxes.reserveCapacity(min(results.count, numItemsThreshold))
-    for prediction in results.prefix(numItemsThreshold) {
-      guard let topLabel = prediction.labels.first else { continue }
-      let invertedBox = CGRect(
-        x: prediction.boundingBox.minX, y: 1 - prediction.boundingBox.maxY,
-        width: prediction.boundingBox.width, height: prediction.boundingBox.height)
-      let imageRect = VNImageRectForNormalizedRect(
-        invertedBox, Int(inputSize.width), Int(inputSize.height))
-      let label = topLabel.identifier
-      boxes.append(
-        Box(
-          index: labels.firstIndex(of: label) ?? 0, cls: label, conf: topLabel.confidence,
-          xywh: imageRect, xywhn: invertedBox))
-    }
-    return boxes
-  }
-
-  // MARK: - Raw tensor processing (NMS-free YOLO26 + traditional)
-
-  /// Dispatches a raw `MLMultiArray` tensor to the appropriate decoder based on output shape (end2end vs traditional).
-  ///
-  /// - Parameter prediction: The raw MLMultiArray output from the model.
-  /// - Returns: An array of detected boxes.
-  private func processRawResults(_ prediction: MLMultiArray) -> [Box] {
-    let shape = prediction.shape.map { $0.intValue }
-    guard shape.count == 3 else { return [] }
-    let strides = prediction.strides.map { $0.intValue }
-    let pointer = prediction.dataPointer.assumingMemoryBound(to: Float.self)
-    let confThreshold = Float(confidenceThreshold)
-
-    // Detect format: end2end [1, max_det, 6] vs traditional [1, 4+nc, num_anchors]
-    let isEnd2End = shape[2] < shape[1]
-    if isEnd2End {
-      return processEnd2EndResults(
-        pointer: pointer, shape: shape, strides: strides, confThreshold: confThreshold)
-    } else {
-      return processTraditionalResults(
-        pointer: pointer, shape: shape, strides: strides, confThreshold: confThreshold)
-    }
-  }
-
-  /// Processes YOLO26 end2end output: [1, max_det, 6] = [x1, y1, x2, y2, conf, class_id] (xyxy pixel coords).
-  ///
-  /// - Parameters:
-  ///   - pointer: Pointer to the raw float data.
-  ///   - shape: The tensor shape [1, max_det, 6].
-  ///   - strides: The tensor strides for correct indexing.
-  ///   - confThreshold: Minimum confidence to include a detection.
-  /// - Returns: An array of detected boxes.
-  private func processEnd2EndResults(
-    pointer: UnsafeMutablePointer<Float>, shape: [Int], strides: [Int],
-    confThreshold: Float
-  ) -> [Box] {
-    let numDetections = shape[1]
-    let numFields = shape[2]
-    let detStride = strides[1]
-    let fieldStride = strides[2]
-    var boxes = [Box]()
-
-    for i in 0..<numDetections {
-      let base = i * detStride
-      let conf = pointer[base + 4 * fieldStride]
-      guard conf > confThreshold else { continue }
-
-      let x1 = CGFloat(pointer[base])
-      let y1 = CGFloat(pointer[base + fieldStride])
-      let x2 = CGFloat(pointer[base + 2 * fieldStride])
-      let y2 = CGFloat(pointer[base + 3 * fieldStride])
-      let classIndex = numFields > 5 ? Int(pointer[base + 5 * fieldStride]) : 0
-
-      let imageRect = inputRect(
-        fromModelRect: CGRect(x: x1, y: y1, width: x2 - x1, height: y2 - y1))
-      let normalizedBox = normalizedRect(fromInputRect: imageRect)
-      let label = classIndex < labels.count ? labels[classIndex] : "\(classIndex)"
-
-      boxes.append(
-        Box(index: classIndex, cls: label, conf: conf, xywh: imageRect, xywhn: normalizedBox))
-      if boxes.count >= numItemsThreshold { break }
-    }
-    return boxes
-  }
-
-  /// Processes traditional YOLO output: [1, 4+nc, num_anchors] in xywh format, requiring Swift NMS.
-  ///
-  /// - Parameters:
-  ///   - pointer: Pointer to the raw float data.
-  ///   - shape: The tensor shape [1, 4+nc, num_anchors].
-  ///   - strides: The tensor strides for correct indexing.
-  ///   - confThreshold: Minimum confidence to include a detection.
-  /// - Returns: An array of detected boxes after non-maximum suppression.
-  private func processTraditionalResults(
-    pointer: UnsafeMutablePointer<Float>, shape: [Int], strides: [Int],
-    confThreshold: Float
-  ) -> [Box] {
-    let numFeatures = shape[1]
-    let numAnchors = shape[2]
-    let numClasses = numFeatures - 4
-    let iouThresh = Float(iouThreshold)
-    let featureStride = strides[1]
-    let anchorStride = strides[2]
-
-    var candidateBoxes = [CGRect]()
-    var candidateScores = [Float]()
-    var candidateClasses = [Int]()
-
-    for j in 0..<numAnchors {
-      let anchorOffset = j * anchorStride
-      var bestScore: Float = 0
-      var bestClass = 0
-      for c in 0..<numClasses {
-        let score = pointer[(4 + c) * featureStride + anchorOffset]
-        if score > bestScore {
-          bestScore = score
-          bestClass = c
-        }
-      }
-      guard bestScore > confThreshold else { continue }
-
-      let x = pointer[anchorOffset]
-      let y = pointer[featureStride + anchorOffset]
-      let w = pointer[2 * featureStride + anchorOffset]
-      let h = pointer[3 * featureStride + anchorOffset]
-      candidateBoxes.append(
-        CGRect(x: CGFloat(x - w / 2), y: CGFloat(y - h / 2), width: CGFloat(w), height: CGFloat(h))
-      )
-      candidateScores.append(bestScore)
-      candidateClasses.append(bestClass)
-    }
-
-    let selectedIndices = nonMaxSuppression(
-      boxes: candidateBoxes, scores: candidateScores, threshold: iouThresh)
-
-    var boxes = [Box]()
-    for i in selectedIndices.prefix(numItemsThreshold) {
-      let rect = candidateBoxes[i]
-      let imageRect = inputRect(fromModelRect: rect)
-      let normalizedBox = normalizedRect(fromInputRect: imageRect)
-      let classIndex = candidateClasses[i]
-      let label = classIndex < labels.count ? labels[classIndex] : "\(classIndex)"
-      boxes.append(
-        Box(
-          index: classIndex, cls: label, conf: candidateScores[i], xywh: imageRect,
-          xywhn: normalizedBox))
-    }
-    return boxes
-  }
-}
