+++ ios/Classes/SwiftYOLOPlatformView.swift
@@ -5,6 +5,7 @@ import AVFoundation
 // Swift 6 strict concurrency.
 @preconcurrency import Flutter
 import UIKit
+import YOLO
 
 // Helper extension for Float to Double conversion
 extension Float {
